able to change the code

+++ frontend/src/pages/persons/index.person.tsx
@@ -7,15 +7,20 @@ import {
   Row,
   Table,
   Button,
+  Card,
+  Image,
 } from 'react-bootstrap';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
 import moment from 'moment';
+import Swal from 'sweetalert2';
 
+const BASE_URL = 'https://api.github.com/users';
 const PersonIndex: React.FC = () => {
   const [persons, setPersons] = useState<any[]>([]);
   const [filter, setFilter] = useState('');
   const [sort, setSort] = useState('asc');
+  const [profile, setProfile] = useState<any>({});
 
   const fetchPerson = async () => {
     try {
@@ -34,10 +39,17 @@ const PersonIndex: React.FC = () => {
   }, []);
 
   async function handleDelete(id: number) {
-    if (window.confirm('Are you sure you want to delete this record?')) {
-      await axios.delete(`/person/${id}`);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'Deleting this record will not be undo!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it',
+    }).then(async (result) => {
+      if (result.isConfirmed) await axios.delete(`/person/${id}`);
       await fetchPerson();
-    }
+    });
   }
 
   async function handleChangeFilter(event: ChangeEvent<HTMLSelectElement>) {
@@ -48,11 +60,16 @@ const PersonIndex: React.FC = () => {
     setSort(event.target.value);
   }
 
+  async function handleFetchGithubProfile(username: string) {
+    const { data }: { data: any } = await axios.get(`${BASE_URL}/${username}`);
+    setProfile(data);
+  }
+
   return (
     <Container>
       <Row>
         <Col lg={12}>
-          <h2>Person Page</h2>
+          <h2>Developer Page</h2>
           <div className='d-flex justify-content-between mb-3'>
             <Link
               to='/create'
@@ -81,7 +98,9 @@ const PersonIndex: React.FC = () => {
                 <option value='firstName'>Firstname</option>
                 <option value='lastName'>Lastname</option>
               </Form.Select>
-              <Button onClick={fetchPerson}>Change</Button>
+              <Button className='btn btn-info' onClick={fetchPerson}>
+                Apply
+              </Button>
             </div>
           </div>
 
@@ -109,19 +128,27 @@ const PersonIndex: React.FC = () => {
                     <img
                       src={person.avatar_url}
                       alt={person.firstName}
-                      width='50'
+                      style={{ width: 32, height: 32, objectFit: 'cover' }}
                     />
                   </td>
                   <td>{person.firstName}</td>
                   <td>{person.lastName}</td>
-                  <td>{moment(person.bornDate).format('MMMM Do YYYY')}</td>
+                  <td>{person.bornDate}</td>
                   <td>{person.height}</td>
                   <td>{person.spouse}</td>
                   <td>{person.motherName}</td>
                   <td>{person.fatherName}</td>
-                  <td>{person.createdAt}</td>
+                  <td>{moment(person.createdAt).fromNow()}</td>
                   <td>
                     <ButtonGroup>
+                      <button
+                        className='btn btn-dark btn-sm'
+                        onClick={() =>
+                          handleFetchGithubProfile(person.username)
+                        }
+                      >
+                        Github Info
+                      </button>
                       <Link
                         to={`/${person._id}/edit`}
                         className='btn btn-info btn-sm'
@@ -140,6 +167,23 @@ const PersonIndex: React.FC = () => {
               ))}
             </tbody>
           </Table>
+          {profile && (
+            <Card>
+              <Card.Body>
+                <Image src={profile.avatar_url} alt={profile.name} /> <br />
+                <p>
+                  Name: {profile.name} <br />
+                  Biography: {profile.bio} <br />
+                  Blog: {profile.blog} <br />
+                  Company: {profile.company} <br />
+                  followers: {profile.followers} <br />
+                  following: {profile.following} <br />
+                  Created: {moment(profile.created_at).fromNow()} <br />
+                </p>
+                <a href={`https://github.com/${profile.login}`}>View Profile</a>
+              </Card.Body>
+            </Card>
+          )}
         </Col>
       </Row>
     </Container>
